ai_intelligence_layer/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
httpx==0.27.2
google-generativeai==0.8.3
python-dotenv==1.0.1
